+++ src/libtrx/include/libtrx/config/types_tr2.h
@@ -77,8 +77,8 @@ typedef struct {
     } ui;
 
     struct {
-        int32_t sound_volume;
-        int32_t music_volume;
+        float sound_volume;
+        float music_volume;
         bool enable_lara_mic;
         float underwater_ambient_volume;
         float underwater_music_volume;
